requirements.txt
autopep8==2.0.1
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycodestyle==2.10.0
pymongo==4.3.3
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
